research/mlperf_reinforcement/tensorflow/minigo/requirements.txt
argh
autopep8>=1.3
google.cloud.logging
numpy>=1.14.0
petname
protobuf
pygtp>=0.4
pylint
sgf==0.5
six
tqdm>=4.17

httplib2
google-cloud-storage
google-auth
oauth2client==4.1
mlperf_compliance==0.0.6
